+++ socks5/NKLIProxyClient/DeDupeProxy/Titanium/TitaniumController.cs
@@ -59,10 +59,10 @@ namespace NKLI.DeDupeProxy
 
         static bool DebugDedupe = false;
         static bool DebugSql = false;
-        static int NumObjects;
-        static int NumChunks;
-        static long LogicalBytes;
-        static long PhysicalBytes;
+        static ulong NumObjects;
+        static ulong NumChunks;
+        static ulong LogicalBytes;
+        static ulong PhysicalBytes;
         static decimal DedupeRatioX;
         static decimal DedupeRatioPercent;
         public static int maxObjectSizeHTTP;
